Reversing order of union types for v

+++ python_variable_annotations/7-to_kv.py
@@ -5,7 +5,7 @@ Module for taking a mixed set of types and returning a tuple.
 from typing import Union
 
 
-def to_kv(k: str, v: Union[float, int]) -> tuple:
+def to_kv(k: str, v: Union[int, float]) -> tuple:
     """
     Type-annotated function that takes a string an an int or float and
     returns a tuple. The first element of the tuple is a string and
